// school-udemy/src/generated-sources/api/model/alumno.ts
/**
 * Generated by orval v7.2.0 🍺
 * Do not edit manually.
 * WebApi
 * OpenAPI spec version: 1.0
 */
import type { Matricula } from './matricula';

export interface Alumno {
  /** @nullable */
  direccion?: string | null;
  /** @nullable */
  dni?: string | null;
  edad?: number;
  /** @nullable */
  email?: string | null;
  id?: number;
  /** @nullable */
  matriculas?: Matricula[] | null;
  /** @nullable */
  nombre?: string | null;
}